Atualiza nova organização de arquivos
