--- Classificador de Acessos/__init__.py
@@ -1,3 +1,0 @@
-from class_buy import carrega_acessos
-
-X, Y = carrega_acessos()
