+++ Classificador de Acessos/class_busc.py
@@ -1,0 +1,1 @@
+from classes import carrega_busca
